change widget coordinate system, to use float value from 0 to 1 instead of int from 0 to 100

+++ include/common.h
@@ -8,10 +8,10 @@
 extern WindowState_t currentWindowState;
 
 typedef struct CoomonPos{
-    int spacingTop;
-    int spacingLeft;
-    int width;
-    int height;
+    float spacingTop;
+    float spacingLeft;
+    float width;
+    float height;
 } CommonPos_t;
 
 #endif
